QmlDesigner: Fix picking of Repeater3D/Loader3D generated nodes
There is no top level Repeater3D node with an actual instance
when the Repeater3D is inside a component. The pick target was not
correctly set in this case.

Now we set pick target on every encountered Repeater3D object so
we can propagate it to every object they create.

The same was done for Loader3D, which also creates nodes dynamically.

Fixes: QDS-5745
Change-Id: I537de28ff7426adaa29c2febc6c4b3416baeccb5
Reviewed-by: Samuel Ghinet <samuel.ghinet@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>