macOS: Add tool for decoupling Qt Creator permissions from user applications
The TCC (Transparency, Consent, and Control) subsystem on macOS handles
access to private private data such as the Photos library, Calendar,
Reminders, or the ability to record the screen (among other things).

When an application is launched by launchd, for example by starting it
from Finder, it becomes the "responsible" process for permissions, and
any child processes launched by the application will both inherit the
application's permissions, and any permission requests from the sub-
processes will be requested on behalf of the application.

This is not ideal when Qt Creator is running user applications, as
we want these applications to properly request and manage their own
permissions, instead of inheriting and relying on the permissions
that Qt Creator may (or may not) have.

To fix this we use a small helper binary that explicitly disclaims
any responsibilities from the parent process, before exec'ing the
user application. We limit the usage of this helper to user apps,
as build tools and other helpers used by Qt Creator should still
be attributed to Qt Creator.

The functionality could have been implemented as a toggleable
feature in processlauncher, but since that helper uses QProcess,
which internally doesn't use posix_spawn on macOS, we would need
some heavy refactoring of QProcess or processlauncher to support
this feature. Keeping it contained in a standalone tool, that can
also be used for testing outside of Qt Creator, makes more sense
in light of this.

With the helper tool in place we can then remove a long list of
permission usage descriptions from Qt Creator's Info.plist that
were only there to allow user applications to not have to add
their own descriptions. Removing these makes it more predicable
for user applications that they need to manage their own
permissions.

It also stops Qt Creator from requesting permissions for Contacts,
Calendars, Photos, etc, at first startup, which looks weird if
you don't expect or want your IDE to manage your reminders or
photo libraries.

[ChangeLog][macOS] Applications launched from Qt Creator will no
longer inherit hardened runtime entitlements from Qt Creator, which
may result in the application needing to declare its own set of
entitlements if the hardened runtime is enabled.

Fixes: QTCREATORBUG-26743
Change-Id: I980d0ed98d875f0f3ec63a54f91f1d9fc7775062
Reviewed-by: Eike Ziller <eike.ziller@qt.io>