Install LocalAddressFinder also as stop dependency
The LocalAddressFinder was installed as a start dependency
for MemcheckToolRunner. However, it wasn't registered as
stop dependency for it. This means that when
MemcheckToolRunner was finished, its RunControl couldn't
finish as one of its RunWorkers wasn't finished yet.

Install LocalAddressFinder as a stop dependency and
implement a stop() method which disconnects the conneciton.

This also prevents the ApplicationOutput pane to open
a new tab on successive run of Memcheck for the same
application.

Fixes: QTCREATORBUG-26758
Change-Id: Ib71ddbe787e5c1bcc7b96687ebf15c9b63b714dc
Reviewed-by: hjk <hjk@qt.io>