CMakePM: Allow "Force to ..." context menu entries also for keys
Previously the "Force to ..." context menu entries were enabled
only for the items on the second column (values).

There is no reason not to have the same behavior for the items
on the first column (keys).

This way the user doesn't have the impression that something is
wrong if they trigger the context menu on the key and not on
the value in the list.

Change-Id: I02fcaf25340f2dbcbb808be8f27047acaaee5471
Reviewed-by: Eike Ziller <eike.ziller@qt.io>