Fix processlauncher for tests
The socket/name for QLocalServer must be a path that is short enough for
unix sockets (<= 104 chars).

QLocalServer already ensures that the "name" is an absolute path, by
prepending QDir::tempPath() if it isn't. For Qt Creator that worked, but
the temporary directory set up for the unittests resulted in a too long
combined path.

Also, we should avoid directly using QDir::tempPath() and use
Utils::TemporaryDirectory to keep all our temporary files at one place.

So, make the "name" explicitly an absolute path in our specific
temporary directory, and make the name shorter for the test.

See 6feed6e656f250b8783f80341e81a5c87ec7ef6c for a similar issue.

Change-Id: I07dc4233db1c9d353cd6977da43850256057ba55
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: hjk <hjk@qt.io>