ActionManager: Update proxy action more directly
The proxy action was updating from the actual action delayed by a
QueuedConnection. This was introduced in
c3772bfd4c52c453179b7e121900c06f72ed36a2 to avoid disconnecting and
reconnecting to QAction::changed while handling that same
QAction::changed.

This introduces a delay between the actual action's change and when it
is reflected in the proxy action.

ProxyAction::update only needs to disconnect signals that would trigger
because of the changes done to the ProxyAction. So it doesn't need to
disconnect/connect from the actual Action's changed signal. Only
disconnect/connect the signals that we really need, and get rid of the
QueuedConnection again.

Fixes: QTCREATORBUG-26363
Change-Id: I2b851af35fcb4a3931b645896dd77477cee79812
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: hjk <hjk@qt.io>