WebAssembly: Fix launching of target.html with qmake-based projects
A recent fix for launching CMake-based projects (where target name
differs from project name) broke launching of any qmake-based project.

The reason for the failure is that the code uses to the buildKey in
order to determine the html file name. CMakeProjectManager holds a
buildKey that corresponts to the target name (which is the right base
name for the html file), but qmake holds the project file path as
buildKey.

This change causes the targetFilePath to be used to determine the html
filee name. It works with CMake and qmake.

Amends: b1db3d2506a9badddb149d876a28ed1f61bd27cc

Fixes: QTCREATORBUG-26849
Change-Id: Ifaa8a279317d7fdb54ceea8198a374a317d9c2d7
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>