Remove "iconlister" tool
It was created in an initiative by a corporate design team, in
order to gather an overview of and coordinate creation of ui
elements. No-one from that team is either still there there or
still pursues this topic.

Furthermore, the Qt Creator repository went back to the mode of
adding binary assets from various sources outside the repository.

Change-Id: I1bd4cbe232d2b2eae8181e3c6e10468266f3f05a
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>