UpdateInfo: (Optionally) Inform user about new Qt versions
New Qt patch or minor releases are not considered "updates" by the
installer framework. The Qt packages are set up this way, to allow
installations of multiple patch versions of the same minor Qt versions.

Explicitly check if a newer Qt version than the latest installed one was
released since we last checked for updates, and inform the user with a
link to the Qt blog, and a button to start the package manager. We don't
show anything when checking for updates the very first time, since at
that point we must assume that the user installed the newest Qt version
they want.

Fixes: QTCREATORBUG-26708
Change-Id: I68374b3242b64d392a16de1abaac1292e0d534cf
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>