python: update docs related to using ui files
This changes the tutorial described in the Creator docs
for the UI-based Python application.

The main reason is that loading an ui file is not recommended,
and discouraged over using uic tools to generate Python code
from ui files.

Some changes to code styling were included to the file as well,
mainly the exec_() to exec() changes.

Change-Id: Ibe1f4d62d2252edecd5f3475250f1287d8213faa
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>