Utils: Optimize SmallString allocation
QStringEncoder returns a much higher size than normally need. Because
allocations are not cheap the string is now copied to an temporary array
on the stack if the required size is propably smaller than the small string
optimization. The size is then computed from the transcoded string and
used for the allocation.

Change-Id: I1e1ef40094ad3040557bc5dd098a0b89006174c5
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>