StudioWelcome: Avoid opening two projects on double click
This issue was introdced because we do not go to the edit mode after
opening a project. We stay in the WelcomeMode until the project is
loaded. Since the recent project list did change a second click
(aka double click) will open a second project.

We now block the update of the model for 1000ms and also block opening another
project for the same duration.

Task-number: QDS-6286
Change-Id: I52ed6aa3686594b63fe4a6134c6db1287bca1dcb
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>