QmlDesigner: Fix creation date issue on macOS
On macOS the creation date was not according to the date/time when the
example was unzipped it was rather the date/time when the example folder
was zipped. By removing the already existing folder and create it again
to hold the content of the zip archive this issue should b circumvented.

Task-number: QDS-6288
Change-Id: Icf18c4fcddcacf69feeca536f6d0e81f4c126afe
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>