Mimetypes v3: Re-introduce regexp as MIME magic option
Re-implements d8055b62f420f244728aae4069cc2b0a5983afe4

This extends the freedesktop standard, which really is lacking in many
ways.
Adds a type "regexp", which interprets the "value" as a
QRegularExpression with MultilineOption and DotMatchesEverythingOption.
The range/offset is the range of the sub-string that is matched against
the regular expression. If the value for the end of the range is
omitted, or the same as the start, the whole string is matched against
the regular expression. The mask is ignored.

Can be chosen as the type when editing MIME magic in the MIME type settings.

Change-Id: I1690d43a83848ec0cf8aeea5be20b7cde7cdc27b
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>