Remove "Device" keyword from platform device name
QC uses the display name set in IDeviceFactory to refer to the
platform name. However, the display name usually have the "Device"
keyword at the end because it's coming form the IDeviceFactory.
That would work fine for example in new device creation dialog.
But QC is using that in other places like new file/project wizard
to show the supported platform for a project, and that's where
having "Device" seems unnecessary or wrong.

This fixes that behavior, but removing that keyword when calling
KitFeatureProvider::displayNameForPlatform().

Change-Id: I5fd4bbcb1ec2579f8b7e86226a7d50b6d7807382
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>