DirectoryFilter: Fix missing file entries before first refresh
After re-starting Qt Creator, the filter was reading the file path cache
from the settings, but it didn't set the corresponding file iterator
that is used when actually invoking Locator.
That lead to the files only being available after the first (automatic)
refresh after a few minutes of Qt Creator running (depending on
settings).

Fixes: QTCREATORBUG-26381
Change-Id: I819961ad3b146d339e760f6648b8b01f437040f7
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: David Schulz <david.schulz@qt.io>