Android: fix logcat retrieval on pre Android Nougat
Logcat on pre Android Nougat (API <= 23) was broken because the the
returned output from the adb command had lots of null characters \u0000
in the string, and thus the comparison against the deployed package
name was failing and the pid was always -1. This makes sure to remove
any null characters before the comparison and simplifies the code a bit.

Task-number: QTBUG-101673
Fixes: QTCREATORBUG-26732
Change-Id: I3f185eceab376dbb0f4f11cf628b71e058ee5583
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>