QmlDesigner: Fix font preview tooltip image
The preview tooltip was recently reduced in size, but font previews
were still being rendered the old size and scaled down, degrading
image quality. The new default size is bit too small to render the
sample text of fonts nicely, so added an option to set unscaled image
to the tooltip, allowing us to use twice as wide preview images for
fonts.

Fixes: QDS-6486
Change-Id: Ieaabfbea11e47509de7cd6aed93464d8595ea541
Reviewed-by: Samuel Ghinet <samuel.ghinet@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>